requirements.txt
halo==0.0.31
jsonpath_ng==1.5.3
Pillow==9.1.1
python-dotenv==0.20.0
requests==2.28.1
riotwatcher==3.2.3
